Merge in ASDF/ese_parser from ASDF-3971_2 to ASDF-3971

* commit '39c6ef6b97cadbc9df8c2a8b8e0d6487699301f1':
  Remove unused build script
  Improve documentation
  fixup docker files (without clobbering the +x)
  Revert "fixup docker builds"
  fixup docker builds
  Fix build
  Fix up windows Jenkinsfile
  Add missing file
  Clippy cleanup
  Small cleanup
  Refactor ese_parser to avoid pulling esent code into the Windows library. Moved ese_parser from app to examples and set it up as a proper example within the libs cargo.toml. Added 'nt_comparison' feature to turn on esent inclusion rather than just tying it to a Windows compilation.
  Avoid println in lib code

+++ lib/src/vartime.rs
@@ -129,7 +129,7 @@ pub fn get_date_time_from_filetime(filetime: u64) -> DateTime<Utc> {
 }
 
 pub fn VariantTimeToSystemTime(dateIn: f64, st: &mut SYSTEMTIME) -> bool {
-	if dateIn < 1 as f64 || dateIn <= (DATE_MIN as f64 - 1.0) || dateIn >= (DATE_MAX as f64 + 1.0) {
+	if dateIn < 1_f64 || dateIn <= (DATE_MIN as f64 - 1.0) || dateIn >= (DATE_MAX as f64 + 1.0) {
 		return false;
     }
 
@@ -198,7 +198,7 @@ pub fn VariantTimeToSystemTime(dateIn: f64, st: &mut SYSTEMTIME) -> bool {
 
 #[test]
 fn test_vartimes() {
-	let t1 : f64 = 44_286.466_608_796_3; 
+	let t1 : f64 = 44_286.466_608_796_3;
 	let mut st : SYSTEMTIME = unsafe { std::mem::MaybeUninit::<SYSTEMTIME>::zeroed().assume_init() };
 	VariantTimeToSystemTime(t1, &mut st);
 	assert_eq!(st.wYear, 2021);
@@ -211,7 +211,7 @@ fn test_vartimes() {
 	assert_eq!(st.wMilliseconds, 0);
 }
 
-#[cfg(target_os = "windows")]
+#[cfg(all(feature = "nt_comparison", target_os = "windows"))]
 #[test]
 fn test_curr_time_with_API() {
 	use std::mem::MaybeUninit;
